+++ gui.py
@@ -16,7 +16,7 @@ Usage:
     python gui.py
 """
 
-__version__ = "1.4.2"
+__version__ = "1.4.3"
 __author__ = "Finlay Crawley"
 
 import sys